+++ app/controllers/users/omniauth_callbacks_controller.rb
@@ -40,8 +40,15 @@ module Users
                                                       attrs: auth,
                                                       identifiable: current_user)
 
-        flash[:notice] = _('Linked succesfully')
+        flash[:notice] = _('Linked successfully')
+
         redirect_to root_path
+      elsif user.id != current_user.id
+        # If a user was found but does NOT match the current user then the identifier has
+        # already been attached to another account (likely the user has 2 accounts)
+        flash[:alert] = format(_('The current %{description} iD has been already linked to a user with email %{email}'),
+                               description: identifier_scheme.description, email: user.email)
+        redirect_to edit_user_registration_path
       end
     end
     # rubocop:enable Metrics/AbcSize, Metrics/MethodLength